Add support for a couple of basic commands.
Now JDS is able to jump to a new directory, and return the current dirlist.

We have also added functions for a few other basic commands, like moving
forward and backward in a dirlist, but they are just stubs for now; their
implementation will come next.

And there is no "generic response" anymore.  At least we don't use a
predefined struct for that.  And in fact we don't need to: it's probably
better to let the returned JSON vary for each command, as needed.   Most
commands will just return the current state of the dirlist after doing
their work.