Standardize error responses per RFC 7807.
We now return a Problem Details structure, as detailed in RFC 7807, for
every error response.

I wasn't aware of this convention earlier; this was suggested by Google
Gemini when I asked it if there are any standard conventions for returning
error responses from HTTP services (that typically return JSON responses).